add missing cast

+++ jdk17u0117/src/hotspot/share/jtsan/jtsanGlobals.hpp
@@ -7,6 +7,6 @@ bool is_jtsan_initialized(void);
 void set_jtsan_initialized(bool value);
 
 void increment_gc_epoch(void);
-uint8_t get_gc_epoch(void);
+uint16_t get_gc_epoch(void);
 
 #endif
